move get_id_command to abstractprotocol
